@@ -1,57 +1,36 @@
-package com.chad.baserecyclerviewadapterhelper.activity.emptyview.adapter;
+package com.chad.baserecyclerviewadapterhelper.activity.emptyview.adapter
 
-import android.content.Context;
-import android.view.LayoutInflater;
-import android.view.ViewGroup;
+import android.content.Context
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.chad.baserecyclerviewadapterhelper.R
+import com.chad.baserecyclerviewadapterhelper.databinding.LayoutAnimationBinding
+import com.chad.baserecyclerviewadapterhelper.entity.Status
+import com.chad.library.adapter.base.BaseQuickAdapter
 
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
+class EmptyViewAdapter : BaseQuickAdapter<Status?, EmptyViewAdapter.VH>() {
 
-import com.chad.baserecyclerviewadapterhelper.R;
-import com.chad.baserecyclerviewadapterhelper.databinding.LayoutAnimationBinding;
-import com.chad.baserecyclerviewadapterhelper.entity.Status;
-import com.chad.library.adapter.base.BaseQuickAdapter;
+    class VH(
+        parent: ViewGroup,
+        val binding: LayoutAnimationBinding = LayoutAnimationBinding.inflate(
+            LayoutInflater.from(parent.context), parent, false
+        ),
+    ) : RecyclerView.ViewHolder(binding.root)
 
-public class EmptyViewAdapter extends BaseQuickAdapter<Status, EmptyViewAdapter.VH> {
-
-
-    static class VH extends RecyclerView.ViewHolder {
-
-        LayoutAnimationBinding binding;
-
-        public VH(@NonNull LayoutAnimationBinding binding) {
-            super(binding.getRoot());
-            this.binding = binding;
-        }
-
-        public VH(@NonNull ViewGroup parent) {
-            this(LayoutAnimationBinding.inflate(LayoutInflater.from(parent.getContext()), parent, false));
-        }
+    override fun onCreateViewHolder(context: Context, parent: ViewGroup, viewType: Int): VH {
+        return VH(parent)
     }
 
-    @NonNull
-    @Override
-    protected VH onCreateViewHolder(@NonNull Context context, @NonNull ViewGroup parent, int viewType) {
-        return new VH(parent);
-    }
-
-    @Override
-    protected void onBindViewHolder(@NonNull VH holder, int position, Status item) {
-        switch (holder.getLayoutPosition() % 3) {
-            case 0:
-                holder.binding.img.setImageResource(R.mipmap.animation_img1);
-                break;
-            case 1:
-                holder.binding.img.setImageResource(R.mipmap.animation_img2);
-                break;
-            case 2:
-                holder.binding.img.setImageResource(R.mipmap.animation_img3);
-                break;
-            default:
-                break;
+     override fun onBindViewHolder(holder: VH, position: Int, item: Status?) {
+        when (holder.layoutPosition % 3) {
+            0 -> holder.binding.img.setImageResource(R.mipmap.animation_img1)
+            1 -> holder.binding.img.setImageResource(R.mipmap.animation_img2)
+            2 -> holder.binding.img.setImageResource(R.mipmap.animation_img3)
+            else -> {}
         }
-        holder.binding.tweetName.setText("Hoteis in Rio de Janeiro");
-        holder.binding.tweetText.setText("O ever youthful,O ever weeping");
+        holder.binding.tweetName.text = "Hoteis in Rio de Janeiro"
+        holder.binding.tweetText.text = "O ever youthful,O ever weeping"
     }
 
 }
