@@ -3,16 +3,20 @@ package com.chad.baserecyclerviewadapterhelper.base;
 import android.os.Build;
 import android.os.Bundle;
 import android.view.View;
-import android.view.ViewGroup;
-import android.view.WindowManager;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import androidx.annotation.LayoutRes;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
+import androidx.core.content.ContextCompat;
+import androidx.viewbinding.ViewBinding;
 
 import com.chad.baserecyclerviewadapterhelper.R;
+import com.chad.baserecyclerviewadapterhelper.databinding.ActivityUniversalRecyclerBinding;
+import com.chad.baserecyclerviewadapterhelper.utils.ExtKt;
 
 /**
  * 文 件 名: BaseActivity
@@ -23,6 +27,17 @@ import com.chad.baserecyclerviewadapterhelper.R;
  * 修改备注：
  */
 public abstract class BaseActivity extends AppCompatActivity {
+
+    public BaseActivity() {
+        this(0);
+    }
+
+    public BaseActivity(@LayoutRes int layoutId) {
+        super(layoutId);
+    }
+
+    protected ViewBinding binding;
+
     /**
      * 日志输出标志getSupportActionBar().
      **/
@@ -65,17 +80,25 @@ public abstract class BaseActivity extends AppCompatActivity {
 
     }
 
+    @Nullable
+    public ViewBinding viewBinding() {
+        return null;
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        // 经测试在代码里直接声明透明状态栏更有效
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-            getWindow().setFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS,
-                    WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS);
+
+        ViewBinding binding = viewBinding();
+        if (binding != null) {
+            setContentView(binding.getRoot());
         }
-        // 这句很关键，注意是调用父类的方法
-        super.setContentView(R.layout.activity_base);
+
+        if (Build.VERSION.SDK_INT >= 21) {
+            getWindow().setStatusBarColor(ContextCompat.getColor(this, R.color.spinner_bg));
+            ExtKt.setStatusBarLightMode(getWindow(), false);
+        }
+
         initToolbar();
     }
 
@@ -95,18 +118,18 @@ public abstract class BaseActivity extends AppCompatActivity {
     }
 
 
-    @Override
-    public void setContentView(int layoutId) {
-        setContentView(View.inflate(this, layoutId, null));
-    }
-
-    @Override
-    public void setContentView(View view) {
-        rootLayout = findViewById(R.id.root_layout);
-        if (rootLayout == null) {
-            return;
-        }
-        rootLayout.addView(view, new ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
-        initToolbar();
-    }
+//    @Override
+//    public void setContentView(int layoutId) {
+//        setContentView(View.inflate(this, layoutId, null));
+//    }
+//
+//    @Override
+//    public void setContentView(View view) {
+//        rootLayout = findViewById(R.id.root_layout);
+//        if (rootLayout == null) {
+//            return;
+//        }
+//        rootLayout.addView(view, new ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
+//        initToolbar();
+//    }
 }
